refactor(connectors): make the WhatsApp page adapter a .js file
It is JavaScript. It runs inside WhatsApp Web's own undeclared module registry,
its globals are the page's rather than Node's, and TypeScript can describe none
of it. Naming it .ts bought nothing and cost a @ts-nocheck, five `declare const`
page-global shims, and a paragraph of comment justifying both.

Dropping all three leaves the file honestly what it is, and makes the
serialisation invariant stricter rather than weaker: with no `declare const`
lines to allow through, the rule collapses to "the module contains nothing at
top level but the exported function". Re-verified by mutation — a hoisted
`const` and a hoisted `function` each still fail it.

Behaviour is unchanged: Function.prototype.toString() serialises a .js export
exactly as it did a .ts one. Re-probed live against WhatsApp Web through the
paired Chrome after the rename — injects and reaches ready on the first probe,
944 chats / 942 messages, same capability set.
